--- launch/panda_pick_and_place.launch.py
@@ -1,36 +1,0 @@
-from ament_index_python.packages import get_package_share_directory
-import launch
-from launch.substitutions import Command, LaunchConfiguration
-from launch.actions import IncludeLaunchDescription
-from launch.launch_description_sources import PythonLaunchDescriptionSource
-import launch_ros
-import os
-
-def generate_launch_description():
-    pkg_share = launch_ros.substitutions.FindPackageShare(package='panda_ros2_gazebo').find('panda_ros2_gazebo')
-    parameter_file_path = os.path.join(pkg_share,
-        "config",
-        "params.yaml"
-    )
-    default_model_path = os.path.join(pkg_share,
-        "description",
-        "models",
-        "panda",
-        "panda.urdf",
-    )
-    
-    panda_pick_and_place_node = launch_ros.actions.Node(
-        package='panda_ros2_gazebo',
-        # namespace=...
-        executable='panda_pick_and_place',
-        name='panda_pick_and_place',
-        parameters=[
-            parameter_file_path,
-            {'share_dir' : pkg_share}
-        ],
-        output='screen'
-    )
-
-    return launch.LaunchDescription([
-        panda_pick_and_place_node
-    ])
