+++ panda_ros2_gazebo/examples/__init__.py
@@ -1,0 +1,1 @@
+from . import scripts
