+++ panda_ros2_gazebo/examples/scripts/models/__init__.py
