+++ panda_ros2_gazebo/examples/scripts/models/panda.py
