+++ panda_ros2_gazebo/examples/scripts/rbd/idyntree/helpers.py
