+++ panda_ros2_gazebo/examples/scripts/rbd/idyntree/inverse_kinematics_nlp.py
