+++ panda_ros2_gazebo/examples/scripts/rbd/utils.py
